% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brd.mat.functions.R
\name{brd.mat.fitness}
\alias{brd.mat.fitness}
\title{brd.mat.fitness allocates offspring to mate pairs in a breeding matrix}
\usage{
brd.mat.fitness(mat = mat, min.fert = 2500, max.fert = 6500, type = "uniform")
}
\arguments{
\item{mat}{is a breeding matrix}

\item{min.fert}{is a integer identifying the minimum number of fertilized eggs a female can produce}

\item{max.fert}{is a integer identifying the maximum number of fertilized eggs a female can produce}

\item{type}{is a character string that can be either "uniform" to uniformly distribution offspring
(see above) across mates or "decline" which non-uniformly distributes offspring across mates}
}
\value{
A matrix with zeros and positive integers representing offspring produced per mate pair
}
\description{
The fecundity for each female in a breeding matrix is randomly drawn from a uniform distribution
representing low and high values for females. Offspring for each female are allocated
uniformly across all of her mates, unless the user changes the type parameter to decline
}
