% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brd.mat.functions.R
\name{brd.mat}
\alias{brd.mat}
\title{brd.mat makes a breeding matrix}
\usage{
brd.mat(moms = 100, dads = 100, lambda.low = 3, lambda.high = 3)
}
\arguments{
\item{moms}{is a integer representing the number of successfully breeding mothers}

\item{dads}{is a integer representing the number of successfully breeding fathers}

\item{lambda.low}{is a integer representing lambda (mean and variance in mating success)}

\item{lambda.high}{is a integer representing lambda (mean and variance in mating success)}
}
\value{
A matrix with zeros and ones
}
\description{
This function makes a matrix where columns identify successfully breeding mothers
and rows identify successfully breeding fathers. Zeros in the matrix identify
mate pairs that did not mate and values of one identify mate pairs that did mate.
Lambda can be set to a range of values to be randomly drawn uniformly from low
and high values. If the user wants lambda to fixed set both the low and high values
to the same number.
}
