% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brd.mat.functions.R
\name{convert2ped}
\alias{convert2ped}
\title{convert2ped converts the data frame from mat.sub.sample into a pedigree}
\usage{
convert2ped(df)
}
\arguments{
\item{df}{is the data frame output from mat.sub.sample}
}
\value{
is a three column pedigree (see above)
}
\description{
This function converts the data frame from mat.sub.sample into a pedigree,
where the pedigree is a three column data frame consisting offspring, mother,
and father IDS and each row represents the parents that produced a given offspring
}
