% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brd.mat.functions.R
\name{mat.stats}
\alias{mat.stats}
\title{mat.sats calculates summary statistics associated with the breeding matrix}
\usage{
mat.stats(mat, id.col = F)
}
\arguments{
\item{mat}{is a breeding matrix}

\item{id.col}{is either TRUE or FALSE to identify if the first column in the matrix
has male IDs}
}
\value{
Returns a data frame with summary statistics associated with the breeding matrix
}
\description{
mat.sats calculates summary statistics associated with the breeding matrix
}
