% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brd.mat.functions.R
\name{mat.sub.sample}
\alias{mat.sub.sample}
\title{mat.sub.sample sub-samples a breeding matrix}
\usage{
mat.sub.sample(mat, noff = 250)
}
\arguments{
\item{mat}{is a breeding matrix where offspring per mate pair have been determined using brd.mat.fitness}

\item{noff}{is the number of offspring randomly sampled}
}
\value{
A data frame with the number of offspring produced per mate pair and the number of offspring sampled described
}
\description{
This function identifies all successfully breeding mate pairs and enumerates how many offspring they produced.
Given the total number of offspring produced in the entire matrix, a per mate pair probability is calculated
and used to randomly sub-sample offspring from each mate pair to some number of sampled offspring defined by
the user.
}
