% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brd.mat.functions.R
\name{ped2mat}
\alias{ped2mat}
\title{ped2mat takes a three-column pedigree and converts it to a breeding matrix}
\usage{
ped2mat(ped)
}
\arguments{
\item{ped}{is a three column data frame consisting offspring, mother, and father IDS
each row represents the parents that produced a given offspring}
}
\value{
Returns a breed matrix
}
\description{
This function takes the output of mat.sub.sample and can convert it
back into a breeding matrix, which can be used to calculate summary statistics
}
